feat(2fa): fix after rebase
ref: MANAGER-14182
