+++ packages/manager/apps/account-disable-2fa/src/components/FileInput/FileInputListItem.tsx
@@ -2,13 +2,13 @@ import React, { FunctionComponent, MouseEvent } from 'react';
 import { ODS_THEME_COLOR_INTENT } from '@ovhcloud/ods-common-theming';
 import { ODS_ICON_NAME, ODS_ICON_SIZE } from '@ovhcloud/ods-components';
 import { OsdsIcon } from '@ovhcloud/ods-components/react';
-import { FileWithError } from './FileInputContainer';
+import { useTranslation } from 'react-i18next';
 
 type FileInputListItemProps = {
-  file: FileWithError;
+  file: File;
   deleteFile: (
     e: MouseEvent<HTMLOsdsIconElement, globalThis.MouseEvent>,
-    file: FileWithError,
+    file: File,
   ) => void;
   accept: string;
   maxSize: number;
@@ -17,7 +17,11 @@ type FileInputListItemProps = {
 export const FileInputListItem: FunctionComponent<FileInputListItemProps> = ({
   file,
   deleteFile,
+  accept,
+  maxSize,
 }) => {
+  const { t } = useTranslation('account-disable-2fa');
+
   const bytesToSize = (bytes: number): string => {
     if (bytes < 1024 * 1024) {
       return `${(bytes / 1024).toFixed(0)} KB`;
@@ -25,7 +29,29 @@ export const FileInputListItem: FunctionComponent<FileInputListItemProps> = ({
     return `${(bytes / (1024 * 1024)).toFixed(2)} MB`;
   };
 
-  const hasError = Boolean(file.errors.length);
+  const parseContentTypes = (inputString: string) => {
+    const typesArray = inputString.split(',').map((type) => type.trim());
+    const extensions = typesArray.map((type) => type.split('/')[1]);
+    const lastType = extensions.pop();
+    return { types: extensions.join(', '), lastType };
+  };
+
+  const errorMessages: string[] = [];
+  if (!accept.includes(file.type)) {
+    const { types, lastType } = parseContentTypes(accept);
+    errorMessages.push(
+      t('account-disable-2fa-file-input-type-file-error', {
+        types,
+        lastType,
+      }),
+    );
+  }
+
+  if (file.size > maxSize) {
+    errorMessages.push(t('account-disable-2fa-file-input-size-file-error'));
+  }
+
+  const hasError = Boolean(errorMessages.length);
 
   return (
     <div
@@ -46,10 +72,8 @@ export const FileInputListItem: FunctionComponent<FileInputListItemProps> = ({
       />
       <p className="text-sm truncate">
         {file.name} {`(${bytesToSize(file.size)})`}
-        {file.errors.map((error, index) => (
-          <span className="block" key={index}>
-            {error}
-          </span>
+        {errorMessages.map((error, index) => (
+          <p key={index}>{error}</p>
         ))}
       </p>
