+++ packages/manager/apps/account-disable-2fa/src/routes/routes.tsx
@@ -30,10 +30,6 @@ export const Routes: any = [
               import('@/pages/create/confirm/Confirm.page'),
             ),
           },
-          {
-            path: '',
-            ...lazyRouteConfig(() => import('@/pages/create/form/Form.page')),
-          },
         ],
       },
       {
