--- packages/manager/apps/account-disable-2fa/src/utils/user.util.ts
@@ -1,10 +1,0 @@
-import { User } from '@/types/user.type';
-
-// const token = ???;
-// const user = decodeJWT(token);
-export const getCurrentUser = (): User => ({
-  legalform: 'other',
-  email: 'monsieur.madame@ovhcloud.com',
-  language: 'fr_FR',
-  subsidiary: 'FR',
-});
