+++ packages/manager/apps/procedures/package.json
@@ -37,6 +37,7 @@
     "i18next-http-backend": "^2.4.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.51.4",
     "react-i18next": "^11.18.1",
     "react-router-dom": "^6.3.0",
     "tailwindcss": "^3.3.3"
