+++ packages/manager/apps/procedures/src/components/FileInput/FileInputContainer.tsx
@@ -6,19 +6,24 @@ import React, {
 } from 'react';
 import { FileInputList } from './FileInputList';
 import { FileInput } from './FileInput';
+import { FileInputTooltip } from './FileInputTooltip';
+
+export type FileWithError = File & {
+  errors: string[];
+};
 
 type Event = {
-  files: File[];
+  files: FileWithError[];
   e: UIEvent | SyntheticEvent;
 };
 
-export type EventHandler = (e: Event) => void | undefined;
+export type FileInputEventHandler = (e: Event) => void | undefined;
 
 export type FileInputProps = {
   id?: string;
-  onChange?: EventHandler;
+  onChange?: FileInputEventHandler;
   multiple?: boolean;
-  value?: File[];
+  value?: FileWithError[];
   accept: string;
   maxSize: number;
   maxFiles: number;
@@ -41,4 +46,5 @@ const Container: FunctionComponent<FileInputProps> = (props) => {
 export const FileInputContainer = Object.assign(Container, {
   FileInput,
   FileList: FileInputList,
+  FileTooltip: FileInputTooltip,
 });
