+++ packages/manager/apps/procedures/src/pages/create/form/FormDocumentFields/documentFieldsConfiguration.json
