+++ packages/manager/apps/procedures/src/public/translations/account-disable-2fa-documents/Messages_fr_FR.json
